+++ client/Packages/com.beamable/Editor/Modules/Content/ContentIO.cs
@@ -86,6 +86,7 @@ namespace Beamable.Editor.Content
 		Promise<Manifest> FetchManifest(string id);
 		IEnumerable<ContentObject> FindAll(ContentQuery query = null);
 		string Checksum(IContentObject content);
+		public LocalContentManifest BuildLocalManifest();
 	}
 
 	[System.Serializable]