+++ client/Packages/com.beamable/Runtime/Modules/Content/IContentCacheFactory.cs
@@ -1,0 +1,35 @@
+using Beamable.Common.Api;
+using Beamable.Coroutines;
+using System;
+
+namespace Beamable.Content
+{
+	public interface IContentCacheFactory
+	{
+		public ContentCache CreateCache(ContentService contentService, string manifestId, Type contentType);
+	}
+
+	public class DefaultContentCacheFactory : IContentCacheFactory
+	{
+		private readonly IHttpRequester _requester;
+		private readonly IBeamableFilesystemAccessor _filesystemAccessor;
+		private readonly CoroutineService _coroutineService;
+
+		public DefaultContentCacheFactory(IHttpRequester requester,
+		                                  IBeamableFilesystemAccessor filesystemAccessor,
+		                                  CoroutineService coroutineService)
+		{
+			_requester = requester;
+			_filesystemAccessor = filesystemAccessor;
+			_coroutineService = coroutineService;
+		}
+		
+		public ContentCache CreateCache(ContentService contentService, string manifestId, Type contentType)
+		{
+			var cacheType = typeof(ContentCache<>).MakeGenericType(contentType);
+			var constructor = cacheType.GetConstructor(new[]
+				                                           {typeof(IHttpRequester), typeof(IBeamableFilesystemAccessor), typeof(ContentService), typeof(CoroutineService)});
+			return (ContentCache)constructor.Invoke(new object[] { _requester, _filesystemAccessor, contentService, _coroutineService });
+		}
+	}
+}