+++ client/Packages/com.beamable/Runtime/Modules/Content/IManifestResolver.cs
@@ -1,0 +1,37 @@
+using Beamable.Api;
+using Beamable.Common;
+using Beamable.Common.Api;
+using Beamable.Common.Content;
+using System.Collections.Generic;
+
+namespace Beamable.Content
+{
+	public interface IManifestResolver
+	{
+		public Promise<ClientManifest> ResolveManifest(IBeamableRequester requester,
+		                                              string url,
+		                                              ManifestSubscription subscription);
+	}
+
+	public class DefaultManifestResolver : IManifestResolver
+	{
+		
+		
+		public Promise<ClientManifest> ResolveManifest(IBeamableRequester requester, string url, ManifestSubscription subscription)
+		{
+			return requester.Request(Method.GET, url, null, true, ClientManifest.ParseCSV, true).Recover(ex =>
+			{
+				// TODO: Put "global" as a constant value somewhere. Currently it lives in a different asm, and its too much trouble.
+				if (ex is PlatformRequesterException err && err.Status == 404 && subscription.ManifestID.Equals("global"))
+				{
+					return new ClientManifest
+					{
+						entries = new List<ClientContentInfo>()
+					};
+				}
+
+				throw ex;
+			});
+		}
+	}
+}